Originated channel is now added to a bridge.
The program checks for an available bridge. If a bridge is available,
then that bridge is reused. If no bridges are active, a new bridge
is created. Once the bridge is created, a channel is originated to
a designated endpoint, and upon answering the call, the channel is
added to the bridge.