Fixed bridge creation and other syntax
Fixed a problem where creating a bridge caused the application to
fail due to bridge being referenced incorrectly for an asynchronous
program. Also fixed other small syntax issues, as well as moving
module.exports to the end of the file so that the functions will not
be created everytime that require is called on the file. Changed the
callback function's name to 'callback'.