Made it possible for inbound callers to dial channels
If the SLA bridge (specified in the dialplan) is numeric,
then create or use a preexisting bridge that has the same
name as the value specified in the dialplan, create a new channel
to dial a specified endpoint.  If the endpoint hangs up, hang up
inbound caller. If the inbound caller hangs up, hang up dialed end.
Application resets when all have hung up.