Allows for dialing of multiple parties.
Added code primarily to the originateChannel method to allow for
multiple hardcoded phones to be dialed upon StasisStart. When one
dialed phone answers, the other dialed phones are hung up. If all
dialed phones hangup or do not answer, the inbound channel is hungup.
Added two more tests for functionality.