Added spacing for readability and a separate update function.
Fixed many alignment issues with indentation and vertical spacing
to make the code more readable. Created a function to perform the
state update instead of creating an update variable each time an
update was required. Modified the errHandler function for the tests
by moving the specific error handlings to the tests that required
them. Removed duplicate tests. Changed the noEndpoints variable
to noStations, which is what was originally intended to be there.