Updated the isFatal function to be simpler.
Updated the isFatal function to check whether an error is non-fatal
and changed the function name to better represent what it is doing.
Also removed a deprecated argument in the .jshintrc file. Also
adjusted the state variables in sla.js to be in an enum format
rather than just stand alone variables.